requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7  # PostgreSQL adapter (optional)

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
python-magic==0.4.27

# Machine Learning and Embeddings
sentence-transformers==2.2.2
faiss-cpu>=1.8.0
numpy==1.24.3
scikit-learn==1.3.0

# LLM Integration
gpt4all>=2.4.0
requests==2.31.0

# Async Processing
celery==5.3.4
redis==5.0.1

# Frontend
streamlit==1.28.1

# Utilities
python-dotenv==1.0.0
Pillow==10.0.1
tqdm==4.66.1

# Development
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.5.2

# Production (optional)
gunicorn==21.2.0
whitenoise==6.6.0